requirements.txt
Flask==3.0.3
flask_sqlalchemy
sqlalchemy_utils
python-dotenv
pymysql
